move CustomModelCheckpoint to new file
